Do not broadcast transaction for a payment request
A transaction from a BIP70 invoice should not be broadcasted by the
wallet.
If a valid PaymentACK is received it is up to the merchant to broadcast
the transaction.
This also fixes a nasty race condition where a transaction was
broadcasted before the merchant had ACK'd it which would likley cause
the user to loose money.

Also updated the memo field from "Electrum" to "Electron Cash".

Fixes #812