Run history_list_filter hook on updated labels
This way an edited label can be filtered directly instead of during
update of the entire history_list.

This can be seen if someone has CashFusion running and the "Hide
CashFusions" option enabled and proceeds to rename a label to start with
"CashFusion ".
Without this fix the transaction is not hidden until the
wallet is restarted or the "Hide CashFusions" option is toggled off and
on.