fusion: Add depth checks
Introduce the notion of 'depth' in fusion transactions.
The user can set a preferred depth that will cause further fusions to
halt once a threshold of all sats (currently hardcoded to 95%) has this
depth.
Only spend coins that satisfies this depth.

By default the depth is set to 0 which means that the fusing will
continue forever (same as current behaviour).