Subscribe to double spend proofs
Add configuration option to enable the double spend detection. Upon a
double spend for an unconfirmed tx in the wallet has occured the
transction will be marked red both in the history list and coins list.

Disable this feature by default since the number of compatible servers
might be an issue currently.